New new iron shot: Bombard
Leda's Unmaking felt a little too unimpactful - Iron Shot with a
little mud spattered on. Try stealing some new tech from Mule and
re-replace it with Bombard, which sends you back a tile when you cast
it.

Might need to nerf the damage a bit more or make the knockback only
a chance (e.g. a coinflip).